Better P4AU support
Moved all duped code into one file.
