+++ palmod/PalMod.vcxproj.filters
@@ -415,9 +415,6 @@
     <Filter Include="Games\Other Platforms\Guilty Gear\GGXX #Reload %28Steam/Community port%29">
       <UniqueIdentifier>{4e591552-6c79-4619-971f-25b5b69225db}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Games\Other Platforms\Persona 4 Arena Ultimax 2.0 %28NESICA%29">
-      <UniqueIdentifier>{7fd4a326-69f3-49f7-884e-98b93bfc1d70}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Games\Capcom Games\Jojos Venture">
       <UniqueIdentifier>{6ba05b08-6fdb-45a1-8d9d-99fc5143b026}</UniqueIdentifier>
     </Filter>
@@ -457,8 +454,14 @@
     <Filter Include="Games\Nintendo\Ultimate Mortal Kombat 3 %28DS%29">
       <UniqueIdentifier>{bf82d014-a0fe-4e32-9ec5-ad44a0a7419f}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Games\Other Platforms\Persona 4 Arena Ultimax %28Steam%29">
+    <Filter Include="Games\Other Platforms\Persona 4 Arena">
+      <UniqueIdentifier>{be4857f1-706e-4cc2-936b-99aa1c526af2}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax %28Steam%29">
       <UniqueIdentifier>{79507b20-b975-408b-ac14-8c08279f9368}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax 2.0 %28NESICA%29">
+      <UniqueIdentifier>{7fd4a326-69f3-49f7-884e-98b93bfc1d70}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -901,7 +904,7 @@
       <Filter>Games\Other Platforms\Guilty Gear\GGXX #Reload %28Steam/Community port%29</Filter>
     </ClCompile>
     <ClCompile Include="Game\Game_P4AU_NESICA.cpp">
-      <Filter>Games\Other Platforms\Persona 4 Arena Ultimax 2.0 %28NESICA%29</Filter>
+      <Filter>Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax 2.0 %28NESICA%29</Filter>
     </ClCompile>
     <ClCompile Include="Game\Game_Venture_A.cpp">
       <Filter>Games\Capcom Games\Jojos Venture</Filter>
@@ -952,7 +955,7 @@
       <Filter>Games\Nintendo\Ultimate Mortal Kombat 3 %28DS%29</Filter>
     </ClCompile>
     <ClCompile Include="Game\Game_P4AU_STEAM.cpp">
-      <Filter>Games\Other Platforms\Persona 4 Arena Ultimax %28Steam%29</Filter>
+      <Filter>Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax %28Steam%29</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1710,10 +1713,10 @@
       <Filter>Games\Other Platforms\Guilty Gear\GGXX #Reload %28Steam/Community port%29</Filter>
     </ClInclude>
     <ClInclude Include="Game\P4AU_NESICA_DEF.h">
-      <Filter>Games\Other Platforms\Persona 4 Arena Ultimax 2.0 %28NESICA%29</Filter>
+      <Filter>Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax 2.0 %28NESICA%29</Filter>
     </ClInclude>
     <ClInclude Include="Game\Game_P4AU_NESICA.h">
-      <Filter>Games\Other Platforms\Persona 4 Arena Ultimax 2.0 %28NESICA%29</Filter>
+      <Filter>Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax 2.0 %28NESICA%29</Filter>
     </ClInclude>
     <ClInclude Include="Game\Game_Venture_A.h">
       <Filter>Games\Capcom Games\Jojos Venture</Filter>
@@ -1803,10 +1806,13 @@
       <Filter>Games\Capcom Games\SFA2_A</Filter>
     </ClInclude>
     <ClInclude Include="Game\Game_P4AU_STEAM.h">
-      <Filter>Games\Other Platforms\Persona 4 Arena Ultimax %28Steam%29</Filter>
+      <Filter>Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax %28Steam%29</Filter>
     </ClInclude>
     <ClInclude Include="Game\P4AU_STEAM_DEF.h">
-      <Filter>Games\Other Platforms\Persona 4 Arena Ultimax %28Steam%29</Filter>
+      <Filter>Games\Other Platforms\Persona 4 Arena\Persona 4 Arena Ultimax %28Steam%29</Filter>
+    </ClInclude>
+    <ClInclude Include="Game\P4AU_Common.h">
+      <Filter>Games\Other Platforms\Persona 4 Arena</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
